+++ lib/mode/page_report_mode.dart
@@ -65,10 +65,16 @@ class PageWidget extends StatefulWidget {
 class PageWidgetState extends State<PageWidget> {
   @override
   Widget build(BuildContext context) {
-    return Builder(
-      builder: (context) => CatcherUtils.isCupertinoAppAncestor(context)
-          ? _buildCupertinoPage()
-          : _buildMaterialPage(),
+    return WillPopScope(
+      onWillPop: () async {
+        widget.pageReportMode.onActionRejected(widget.report);
+        return true;
+      },
+      child: Builder(
+        builder: (context) => CatcherUtils.isCupertinoAppAncestor(context)
+            ? _buildCupertinoPage()
+            : _buildMaterialPage(),
+      ),
     );
   }